refactor: update summary in controller
