+++ hrp/step_websocket.go
@@ -260,16 +260,6 @@ func runStepWebSocket(r *SessionRunner, step *TStep) (stepResult *StepResult, er
 		if err != nil {
 			stepResult.Attachments = err.Error()
 		}
-		// update summary
-		r.summary.Records = append(r.summary.Records, stepResult)
-		r.summary.Stat.Total += 1
-		if stepResult.Success {
-			r.summary.Stat.Successes += 1
-		} else {
-			r.summary.Stat.Failures += 1
-			// update summary result to failed
-			r.summary.Success = false
-		}
 	}()
 
 	// override step variables
@@ -279,8 +269,8 @@ func runStepWebSocket(r *SessionRunner, step *TStep) (stepResult *StepResult, er
 	}
 
 	sessionData := newSessionData()
-	parser := r.GetParser()
-	config := r.GetConfig()
+	parser := r.parser
+	config := r.parsedConfig
 
 	dummyReq := &Request{
 		URL:     step.WebSocket.URL,
